+++ src/main/java/com/io/gittracker/AppConfiguration.java
@@ -1,25 +1,27 @@
 package com.io.gittracker;
 
 import com.io.gittracker.model.PermaStorage;
-import com.io.gittracker.services.TokenService;
+import com.io.gittracker.services.AppStateService;
 import java.io.IOException;
 import org.kohsuke.github.GitHub;
 import org.kohsuke.github.GitHubBuilder;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Lazy;
 
 @Configuration
 public class AppConfiguration {
 
-    private final TokenService tokenService;
+    AppStateService appStateService;
 
-    public AppConfiguration(TokenService tokenService) {
-        this.tokenService = tokenService;
+    public AppConfiguration(@Lazy AppStateService appStateService) {
+        this.appStateService = appStateService;
     }
 
     @Bean
     public GitHub getGithub() throws IOException {
-        return new GitHubBuilder().withOAuthToken(tokenService.getApiKey()).build();
+        String token = appStateService.getAppState().githubToken;
+        return new GitHubBuilder().withOAuthToken(token).build();
     }
 
     @Bean
