+++ src/main/java/com/io/gittracker/model/AppState.java
@@ -15,6 +15,7 @@ public final class AppState implements Serializable {
     private final List<Workspace> workspaces;
     private int currentWorkspaceIndex;
     private int currentGroupIndex;
+    public String githubToken = "EMPTY";
 
     private final Date lastUpdate;
