+++ src/main/java/com/io/gittracker/model/PermaStorage.java
@@ -3,9 +3,6 @@ package com.io.gittracker.model;
 import com.google.gson.*;
 import dev.dirs.ProjectDirectories;
 import java.io.*;
-import java.util.ArrayList;
-import java.util.Date;
-import java.util.List;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
