+++ src/main/java/com/io/gittracker/services/TokenService.java
@@ -1,16 +1,21 @@
 package com.io.gittracker.services;
 
+import com.io.gittracker.model.AppState;
 import org.springframework.stereotype.Service;
 
 @Service
 public class TokenService {
-    private String apiKey;
+    AppState app_state;
+
+    TokenService(AppStateService appStateService) {
+        this.app_state = appStateService.getAppState();
+    }
 
     public void setApiKey(String apiKey) {
-        this.apiKey = apiKey;
+        this.app_state.githubToken = apiKey;
     }
 
     public String getApiKey() {
-        return apiKey;
+        return this.app_state.githubToken;
     }
 }
